move cross backend test to own file
